requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
av==14.0.1
baize==0.22.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
ctranslate2==4.5.0
cycler==0.12.1
datasets==2.18.0
decorator==5.1.1
dill==0.3.8
docker-pycreds==0.4.0
editdistance==0.8.1
einops==0.8.0
einx==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.6
faster-whisper==1.1.1
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.13.1
-e git+https://github.com/fishaudio/fish-speech.git@3d2f842e64c16a58ae3d53779871468ef2977675#egg=fish_speech
flatbuffers==24.12.23
fonttools==4.55.6
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.2.0
funasr==1.1.5
gitdb==4.0.12
GitPython==3.1.44
gradio==5.12.0
gradio_client==1.5.4
grpcio==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.4.2
kaldiio==2.18.0
kiwisolver==1.4.8
kui==1.8.1
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.43.0
loguru==0.7.3
loralib==0.1.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
modelscope==1.17.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
natsort==8.4.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
omegaconf==2.3.0
onnxruntime==1.20.1
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opencc-python-reimplemented==0.1.7
orjson==3.10.15
ormsgpack==1.7.0
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pyarrow-hotfix==0.6
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
pyreadline3==3.5.4
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytorch-wpe==0.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
silero-vad==5.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.0
soxr==0.5.0.post1
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.4.1+cu121
torch-complex==0.4.4
torchaudio==2.4.1+cu121
torchmetrics==1.6.1
torchvision==0.19.1+cu121
tqdm==4.67.1
transformers==4.48.0
triton-windows @ https://github.com/AnyaCoder/fish-speech/releases/download/v0.1.0/triton_windows-0.1.0-py3-none-any.whl#sha256=2cc998638180f37cf5025ab65e48c7f629aa5a369176cfa32177d2bd9aa26a0a
typer==0.15.1
typing_extensions==4.9.0
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.0
vector-quantize-pytorch==1.14.24
wandb==0.19.4
websockets==14.1
Werkzeug==3.1.3
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
